Share a mock server definition between test & dev
The mock server makes it easier to develop and test the frontend in
isolation. The mock server is never used when `NODE_ENV=production`.

A new environment variable, FRONTEND_MOCK_API_SERVER, controls whether
the mock is used (when not in production). This allows the mock to be
disabled when developing the frontend and backend together.