Define <Router> routes manually
React-Static provides `<Routes />`, which can define routes
automatically, based on the same directory structure it uses to build
static HTML pages. The Routes component doesn't play nicely with our
testing setup, however.

For the sake of testing, as well as being explicit, we're now defining
those routes manually, the same as if it *weren't* a statically-rendered
site. This *does* mean that we're now responsible for keeping that
routing in sync with the file structure so that in-app routing aligns
with the working URLs of initial HTTP requests.