Add one more trailing slash in Router path
This doesn't actually affect the behavior of @reach/router.
Unfortunately, it always redirects to a path *lacking* a trailing slash.

Even so, we want to be explicit and consistent about our intention to
use trailing slashes, which are more consistent with paths used for
static content hosting. If we later switch to a different router which
*does* support redirects with trailing slashes, we'll already have the
correct paths defined.