Fix build failure
`react-static` choked on redirecting from `/category/` to `/categories/`
(but only when building, not with the dev server).  I'm not entirely
sure why it was a problem, but removing the redirect fixes it.

In any case, the new behavior actually makes more intuitive sense: A 404
status is now returned for `/category/`.  There is no such page, so the
previous attempt to redirect was just a tiny convenience for an edge
case that would only occur by manually typing the incorrect URL.