Enable a skipped API unit test
The whole API unit test module now uses the "node" Jest environment
instead of the "jsdom" environment.

This test was failing when using the jsdom, but it works in the node
environment and I also verified that it works in the browser, so the
issue is likely something specific to jsdom (or how msw and/or axios
interact with jsdom).