Update components & tests to use the new API...API
Components now use the callbacks and cancellation feature of the new API
interface instead of the old Promise-based one.

One benefit for tests is that they can simply unmount a component rather
than waiting for an AJAX request to complete to avoid a warning when the
given test isn't concerned with the result of the AJAX request.