Swap browser api to useLocation for InstrumentPage
We've been accessing `window.location.pathname` from
src/pages/instrument.tsx, which worked perfectly in the browser.
Unfortunately, there are some circumstances in our jsdom test
environment where `pathname` has the value "/", even when the rest of
the app renders correctly for a different path.

Switching to `useLocation` doesn't affect the app behavior in the
browser, but it fixes the weird behavior in the test environment.