Implement useCategories() hook to fetch/cache data
Categories are stored in the database and they *can* change, but they're
likely to change very infrequently. We *use* the list of categories
frequently on the frontend, so it's valuable to fetch them once per site
load and then reuse the cached category data. ...so that's what the
useCategories() hook does!