Implement baseAuthenticatedRequest() in api.ts
I chose to implement this as a normal function rather than a hook. This
way, the higher level API functions using it can be simple wrappers and
can still be tree-shaken, whereas a hook would have to either return all
authenticated request functions at once or only provide a lower level
API that requires components to specify more implementation details
(URL, method, potentially more).

The downside of this approach is that components making authenticated
requests are responsible for calling useAuth() to get a reference to the
getAccessTokenSilently function, which must then be passed to the
authenticated API function.