Define a base HTML document
We now define a `<title>`, though the same title text is used for every
page at this point. It's easy to update the title for other pages, but
when making that change, we'll want to make sure that the statically
generated files don't set the title to "404" initially for components
that conditionally render the 404 page.